fix(ci): satisfy grok provider readiness guardrails
